replaced header and footer with component-ised NavBar and FooterBar in AboutusPage
