+++ src/Routes.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ListingsPage from "pages/ListingsPage/listingsPage";
-import AboutusPage from "pages/AboutusPage";
+import AboutusPage from "pages/AboutusPage/aboutUsPage";
 import RegisterPage from "pages/RegisterPage";
 import LoginPage from "pages/LoginPage";
 import LandingPage from "pages/LandingPage";
